refactor: Move convert_method_call to its own module
